+++ src/screens/HomeScreen/HomeSrceen.tsx
@@ -21,8 +21,19 @@ const HomeScreen: React.FC<Props> = ({ navigation, route }) => {
   const [profile, setProfile] = useState<UserEntity>();
 
   useEffect(() => {
+    checkAuthToken();
     getProfile();
   }, []);
+  const checkAuthToken = async () => {
+    const auth = await UserLocalStorageService.haveAccessToken().then((res) => {
+      return res;
+    });
+    if (!auth) {
+      navigation.navigate("Auth", {
+        screen: "InputTelNumber",
+      });
+    }
+  };
 
   const getProfile = async () => {
     await VerifiesDataSource.getProfile().then((respone: UserEntity) => {
@@ -72,7 +83,7 @@ const HomeScreen: React.FC<Props> = ({ navigation, route }) => {
             </TouchableOpacity>
             <TouchableOpacity
               onPress={() => {
-                alert("อยู่ระหว่างการพัฒนาระบบ")
+                alert("อยู่ระหว่างการพัฒนาระบบ");
               }}
             >
               <Image
@@ -83,7 +94,7 @@ const HomeScreen: React.FC<Props> = ({ navigation, route }) => {
             </TouchableOpacity>
             <TouchableOpacity
               onPress={() => {
-                alert("อยู่ระหว่างการพัฒนาระบบ")
+                alert("อยู่ระหว่างการพัฒนาระบบ");
               }}
             >
               <Image
@@ -108,7 +119,7 @@ const styles = StyleSheet.create({
   },
   profileWarp: {
     // flex: 0.4,
-    height:"30%",
+    height: "30%",
     backgroundColor: "#4C95FF",
     flexDirection: "row",
     alignItems: "center",